+++ CoronaKurzArbeit/Extensions/TimeBookingExtensions.cs
@@ -32,31 +32,31 @@ namespace CoronaKurzArbeit.Extensions
             return ret;
         }
 
-        public static double GetWorkhours(this DateTime current, KurzarbeitSettingsConfiguration config)
+        public static decimal GetWorkhours(this DateTime current, KurzarbeitSettingsConfiguration config)
         {
             var props = config.GetType().GetProperties();
             foreach (var p in props)
             {
                 if (p.Name == current.DayOfWeek.ToString())
                 {
-                    return Convert.ToDouble(p.GetValue(config) ?? default(double));
+                    return Convert.ToDecimal(p.GetValue(config) ?? default(decimal));
                 }
             }
             return 0;
         }
 
-        public static double GetCoronaWorkhours(this DateTime current, KurzarbeitSettingsConfiguration config)
+        public static decimal GetCoronaWorkhours(this DateTime current, KurzarbeitSettingsConfiguration config)
         {
             var props = config.GetType().GetProperties();
-            var wh = double.MinValue;
+            var wh = decimal.MinValue;
             foreach (var p in props)
             {
                 if (p.Name == current.DayOfWeek.ToString())
                 {
-                    wh = Convert.ToDouble(p.GetValue(config) ?? default(double));
+                    wh = Convert.ToDecimal(p.GetValue(config) ?? default(decimal));
                 }
             }
-            if (wh == double.MinValue) return 0;
+            if (wh == decimal.MinValue) return 0;
 
             return wh * config.CoronaSoll;
         }
